move ListProperties to admin.domain
